disable support for docstring in Python versions below 3.7
